Created logic for card randomization
Created the logic needed for creating a randomized deck and drawing the cards. Renamed first h4 header from 'Spades' to 'Hand' to be used as a simulation for drawing cards. Can draw cards at either 7 or 1 at a time. Can also rerandomize the deck. Added WIP animation and transition for the settings button in top right corner.